+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
-## [0.0.1] - 2026-02-14
+## [0.0.18] - 2026-02-14
 
 ### Added
 - Monorepo migration.
+- Improved websocket client missed event handler, annotated and renamed some options